+++ src/main/java/com/itau/cdc/controller/CategoriaController.java
@@ -16,17 +16,20 @@ import com.itau.cdc.DTO.CategoriaRequest;
 import com.itau.cdc.service.CategoriaService;
 
 @RestController
+//4
 public class CategoriaController {
 
 	@Autowired
+	//1
 	private CategoriaService categoriaService;
 	
 	@PostMapping("/v1/categorias")
 	@Transactional
+	//1
 	public ResponseEntity<?> CriaCategoria(@RequestBody @Valid CategoriaRequest request, UriComponentsBuilder builder){
 		//1
 		Long idCategoria = categoriaService.IncluirCategoria(request);
-		
+		//1
 		URI enderecoConsulta = builder.path("/v1/categorias/{id}").build(idCategoria);
 		
 		return ResponseEntity.created(enderecoConsulta).build();
