+++ src/main/java/com/itau/cdc/controller/CupomConsultaController.java
@@ -1,6 +1,5 @@
 package com.itau.cdc.controller;
 
-import java.net.URI;
 import java.util.Optional;
 
 import javax.transaction.Transactional;
@@ -10,51 +9,38 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import com.itau.cdc.DTO.CupomRequest;
 import com.itau.cdc.entity.Cupom;
-import com.itau.cdc.service.CupomService;
+import com.itau.cdc.service.CupomConsultaService;
 
 @RestController
-public class CupomController {
+//5
+public class CupomConsultaController {
 
 	@Autowired
-	private CupomService novoCupomService;
-	
-	@PostMapping("/v1/cupons")
-	@Transactional
-	private ResponseEntity<?> NovoCupom(@RequestBody @Valid CupomRequest request, UriComponentsBuilder builder){
-		
-		Long idCupom = novoCupomService.NovoCupom(request);
-		
-		URI enderecoConsulta = builder.path("/v1/cupons/{id}").build(idCupom);
-		
-		return ResponseEntity.created(enderecoConsulta).build();
-	}
+	//1
+	private CupomConsultaService cupomConsultaService;
 	
 	@GetMapping("/v1/cupons/{codigo_cupom}")
 	@Transactional
+	//1
 	private ResponseEntity<?> ConsultaCupom(@PathVariable("codigo_cupom") String codigoCupom){
+		//1
+		Optional<Cupom> cupom = cupomConsultaService.ConsultaCupom(codigoCupom);
 		
-		Optional<Cupom> cupom = novoCupomService.ConsultaCupom(codigoCupom);
-		
-		if(cupom.isPresent()) {
-			return ResponseEntity.ok(cupom);
-		}else {
-			return ResponseEntity.notFound().build();
-		}
+		return ResponseEntity.ok(cupom);
 	}
 	
 	@PutMapping("/v1/cupons/{codigo_cupom}")
 	@Transactional
+	//1
 	private ResponseEntity<?> AlteraCupom(@PathVariable("codigo_cupom") String codigoCupom, @RequestBody @Valid CupomRequest request){
-		
-		Cupom cupom = novoCupomService.AlteraCupom(codigoCupom, request);
+		//1
+		Cupom cupom = cupomConsultaService.AlteraCupom(codigoCupom, request);
 		
 		return ResponseEntity.ok(cupom);
