+++ src/main/java/com/itau/cdc/controller/DescricaoLivroController.java
@@ -11,16 +11,19 @@ import com.itau.cdc.DTO.DetalheSiteLivroResponse;
 import com.itau.cdc.service.DescricaoLivroServiceService;
 
 @RestController
+//4
 public class DescricaoLivroController {
 
 	@Autowired
+	//1
 	private DescricaoLivroServiceService descricaoLivroService;
 	
 	@GetMapping("/v1/livros/{id_livro}")
+	//1
 	public ResponseEntity<?> DescricaoLivro(@PathVariable("id_livro") Long idLivro, UriComponentsBuilder builder){
-		
+		//1
 		DetalheSiteLivroResponse livro = descricaoLivroService.DescricaoLivro(idLivro);
-		
+		//1
 		if (livro == null) {
 			return ResponseEntity.notFound().build();
 		}else {
