+++ src/main/java/com/itau/cdc/service/AutorService.java
@@ -8,23 +8,24 @@ import org.springframework.stereotype.Service;
 
 import com.itau.cdc.DTO.AutorRequest;
 import com.itau.cdc.Repository.AutorJpaRepository;
+import com.itau.cdc.configuration.exception.ApiErroException;
 import com.itau.cdc.entity.Autor;
-import com.itau.cdc.exception.ApiErroException;
 
 @Service
+//3
 public class AutorService {
 
-	//1
 	@Autowired
 	private AutorJpaRepository autorJpaRepository;
-	
+	//1
 	public Long IncluirAutor(@Valid AutorRequest request) {
-		
+		//1
 		if((autorJpaRepository.findByEmail(request.getEmail()).isPresent())) {
 			throw new ApiErroException(HttpStatus.UNPROCESSABLE_ENTITY, "Já existe um(a) outro(a) autor(a) com o mesmo email " + request.getEmail());
 		}
 		
 		@Valid
+		//1
 		Autor novoAutor = request.toModel();
 		
 		novoAutor = autorJpaRepository.save(novoAutor);
