+++ src/main/java/com/itau/cdc/service/CupomNovoService.java
@@ -1,7 +1,5 @@
 package com.itau.cdc.service;
 
-import java.util.Optional;
-
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,19 +8,23 @@ import org.springframework.stereotype.Service;
 
 import com.itau.cdc.DTO.CupomRequest;
 import com.itau.cdc.Repository.CupomJpaRepository;
+import com.itau.cdc.configuration.exception.ApiErroException;
 import com.itau.cdc.entity.Cupom;
-import com.itau.cdc.exception.ApiErroException;
 
 @Service
-public class CupomService {
+//3
+public class CupomNovoService {
 
 	@Autowired
 	private CupomJpaRepository cupomJpaRepository;
 
+	//1
 	public Long NovoCupom(@Valid CupomRequest request) {
 
+		//1
 		Cupom novoCupom = request.toModel();
 
+		//1
 		if(cupomJpaRepository.findByCodigo(novoCupom.getCodigo()).isPresent()) {
 			throw new ApiErroException(HttpStatus.UNPROCESSABLE_ENTITY, "Código já cadastrado.");
 		}
@@ -32,27 +34,6 @@ public class CupomService {
 		return novoCupom.getId();
 	}
 
-	public Optional<Cupom> ConsultaCupom(String codigoCupom) {
-
-		return cupomJpaRepository.findByCodigo(codigoCupom);
-
-	}
-
-	public Cupom AlteraCupom(String codigoCupom, @Valid CupomRequest request) {
-		
-		Optional<Cupom> cupomNaBase = cupomJpaRepository.findByCodigo(codigoCupom);
-		
-		if(!cupomNaBase.isPresent()) {
-			return null;
-		}
-		
-		Cupom cupomAlterado = request.toModel();
-		cupomAlterado.AlteraIdCupom(cupomNaBase);
-		
-		cupomJpaRepository.save(cupomAlterado);
-		
-		return cupomAlterado;
-		
-	}
+	
 
 }
