+++ src/main/java/com/itau/cdc/service/EstadoService.java
@@ -10,10 +10,11 @@ import org.springframework.stereotype.Service;
 
 import com.itau.cdc.DTO.EstadoRequest;
 import com.itau.cdc.Repository.EstadoJpaRepository;
+import com.itau.cdc.configuration.exception.ApiErroException;
 import com.itau.cdc.entity.Estado;
-import com.itau.cdc.exception.ApiErroException;
 
 @Service
+//3
 public class EstadoService {
 
 	@Autowired
@@ -22,13 +23,15 @@ public class EstadoService {
 	@PersistenceContext
 	private EntityManager manager;
 	
+	//1
 	public Long IncluirEstado(@Valid EstadoRequest request) {
-		
+		//1
 		if((estadoJpaRepository.findByNome(request.getNome()).isPresent())){
 			throw new ApiErroException(HttpStatus.UNPROCESSABLE_ENTITY, "Já existe um outro estado com o mesmo nome " + request.getNome());
 		}
 			
 		@Valid
+		//1
 		Estado novoEstado = request.toModel(manager);
 		
 		novoEstado = estadoJpaRepository.save(novoEstado);
